Refactor/edicoes impl final

+++ code/backend/moeda/src/main/java/com/lab3/moeda/dto/request/VantagemRequestDTO.java
@@ -5,6 +5,6 @@ public record VantagemRequestDTO(
         String nome,
         String descricao,
         String foto,
-        int custo,
+        short custo,
         int estoque
 ) {}
